made ai computations asynchronous to keep ui responsive moved GameController to GUI namespace switched to c# version 4.5
